feat: Implement InputArea with clear button and split into components
